migrate to pear v2
