+++ index.html
@@ -2,10 +2,6 @@
 <html>
   <body>
     <div id="root"></div>
-    <script
-      type="module"
-      src="./index.js"
-    >
-    </script>
+    <script type="module" src="./src/app.js"></script>
   </body>
 </html>
